+++ src/main/java/com/alexandrialms/Main.java
@@ -4,6 +4,8 @@ import com.alexandrialms.*;
 import com.alexandrialms.dao.*;
 import com.alexandrialms.model.*;
 import com.alexandrialms.util.DBConnection;
+import com.alexandrialms.util.ValidationHelper;
+
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.time.LocalDate;
@@ -16,15 +18,12 @@ public class Main {
     public static void main(String[] args) {
 
 
- UserDAO userDAO = new UserDAO();       
+        List<Book> books = new ArrayList<Book>();
 
-        List<User> users = userDAO.findAll();
-        System.out.println("All users after insert:");
-        for (User u : users) {
-            System.out.println(u.getUserID() + " - " + u.getFirstName() + " " + u.getLastName() +
-                    ", Email: " + u.getEmail() +
-                    ", Role: " + u.getRole() +
-                    ", Active: " + u.isActive());
+        BookDAO bookDAO = new BookDAO();
+        books=bookDAO.findAll();
+        for (Book book : books) {
+            System.out.println(book.getTitle());
         }