+++ src/main/java/com/alexandrialms/dao/BookDAO.java
@@ -1,15 +1,16 @@
 package com.alexandrialms.dao;
 
 import java.sql.*;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.alexandrialms.dao.interfaces.BookDAOInterface;
 import com.alexandrialms.model.Book;
 import com.alexandrialms.util.DBConnection;
+import com.alexandrialms.util.ValidationHelper;
 
-public class BookDAO {
-
+public class BookDAO implements BookDAOInterface {
+    @Override
     public List<Book> findAll() {
         List<Book> books = new ArrayList<>();
         try (Connection conn = DBConnection.getConnection();
@@ -17,14 +18,7 @@ public class BookDAO {
                 ResultSet rs = statement
                         .executeQuery("SELECT book_id, title, isbn, publication_year, category_id FROM books;");) {
             while (rs.next()) {
-                Book book = new Book();
-                book.setBookID(rs.getInt("book_id"));
-                book.setTitle(rs.getString("title"));
-                book.setIsbn(rs.getString("isbn"));
-                book.setPubYear(rs.getInt("publication_year"));
-                book.setCategoryId(rs.getInt("category_id"));
-                books.add(book);
-
+                books.add(mapResultSet(rs));
             }
         } catch (SQLException e) {
             e.printStackTrace();
@@ -32,38 +26,35 @@ public class BookDAO {
         return books;
     }
 
-    public Book findById(int bookID) {
-        Book book = null;
+    @Override
+    public Book findById(Integer bookID) {
+        if (!ValidationHelper.isValidBookID(bookID,this)){
+            return null;
+        }
+        String sql = "SELECT book_id, title, isbn, publication_year, category_id from books where book_id = ?;";
         try (
                 Connection conn = DBConnection.getConnection();
-                PreparedStatement pstm = conn.prepareStatement(
-                        "SELECT book_id, title, isbn, publication_year, category_id from books where book_id = ?;");) {
+                PreparedStatement pstm = conn.prepareStatement(sql);) {
             pstm.setInt(1, bookID);
             ResultSet rs = pstm.executeQuery();
             if (rs.next()) {
-                book = new Book();
-                book.setBookID(rs.getInt("book_id"));
-                book.setTitle(rs.getString("title"));
-                book.setIsbn(rs.getString("isbn"));
-                book.setPubYear(rs.getInt("publication_year"));
-                book.setCategoryId(rs.getInt("category_id"));
+                return mapResultSet(rs);
             }
         } catch (SQLException e) {
             e.printStackTrace();
         }
-        return book;
+        return null;
     }
 
+    @Override
     public boolean insert(Book book) {
-
-        try (Connection conn = DBConnection.getConnection();
-                PreparedStatement pstm = conn.prepareStatement(
-                        "INSERT INTO books (title, isbn, publication_year, category_id) VALUES (?,?,?,?)");) {
+    String sql = "INSERT INTO books (title, isbn, publication_year, category_id) VALUES (?, ?, ?, ?)";
+        try (Connection conn = DBConnection.getConnection();        
+            PreparedStatement pstm = conn.prepareStatement(sql);) {
             pstm.setString(1, book.getTitle());
             pstm.setString(2, book.getIsbn());
             pstm.setInt(3, book.getPubYear());
-            pstm.setInt(4, book.getCategoryId()); // TO DO: crear un método que valide que el category_id está entre 1 y
-                                                  // la categoría máxima
+            pstm.setInt(4, book.getCategoryId()); 
             pstm.executeUpdate();
             return true;
         } catch (SQLException e) {
@@ -72,10 +63,11 @@ public class BookDAO {
         return false;
     }
 
+    @Override
     public boolean update(Book book) {
+        String sql = "UPDATE Book SET title = ?, isbn = ?, publication_year = ?, category_id = ? WHERE book_id = ?;";
         try (Connection conn = DBConnection.getConnection();
-                PreparedStatement pstm = conn.prepareStatement(
-                        "UPDATE Book SET title = ?, isbn = ?, publication_year = ?, category_id = ? WHERE book_id = ?");) {
+                PreparedStatement pstm = conn.prepareStatement(sql);) {
             pstm.setString(1, book.getTitle());
             pstm.setString(2, book.getIsbn());
             pstm.setInt(3, book.getPubYear()); // TO DO: crear un método que valide que el category_id está entre 1 y la
@@ -91,9 +83,14 @@ public class BookDAO {
         return false;
     }
 
-    public boolean delete(int bookID) {
+    @Override
+    public boolean delete(Integer bookID) {
+        if (!ValidationHelper.isValidBookID(bookID, this)){
+            return false;
+        }
+        String sql= "DELETE FROM books WHERE book_id = ?";
         try (Connection conn = DBConnection.getConnection();
-                PreparedStatement pstm = conn.prepareStatement("DELETE FROM books WHERE book_id = ?");) {
+                PreparedStatement pstm = conn.prepareStatement(sql);) {
             pstm.setInt(1, bookID);
             pstm.executeUpdate();
             return true;
@@ -101,6 +98,55 @@ public class BookDAO {
             e.printStackTrace();
         }
         return false;
+    }
+
+    private Book mapResultSet(ResultSet rs) throws SQLException {
+        Book book = new Book();
+        book.setBookID(rs.getInt("book_id"));
+        book.setTitle(rs.getString("title"));
+        book.setIsbn(rs.getString("isbn"));
+        book.setPubYear(rs.getInt("publication_year"));
+        book.setCategoryId(rs.getInt("category_id"));
+
+        return book;
+    }
+
+    @Override
+    public Book findByISBN(String isbn) {
+        if (!ValidationHelper.isValidISBN(isbn)) {
+            return null;
+        }
+        String sql = "SELECT book_id, title, isbn, publication_year, category_id from books where isbn = ?;";
+        try (
+                Connection conn = DBConnection.getConnection();
+                PreparedStatement pstm = conn.prepareStatement(sql);) {
+            pstm.setString(1, isbn);
+            ResultSet rs = pstm.executeQuery();
+            if (rs.next()) {
+                return mapResultSet(rs);
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    @Override
+    public List<Book> findByTitle(String partialTitle) {
+        String sql = "SELECT book_id, title, isbn, publication_year, category_id from books where LOWER(title) LIKE LOWER(?);";
+        List<Book> books = new ArrayList<>();
+        try (
+                Connection conn = DBConnection.getConnection();
+                PreparedStatement pstm = conn.prepareStatement(sql);) {
+            pstm.setString(1, "%" + partialTitle + "%");
+            ResultSet rs = pstm.executeQuery();
+            while (rs.next()) {
+                books.add(mapResultSet(rs));
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return books;
     }
 
 }