+++ src/main/java/com/alexandrialms/dao/interfaces/LoanDAOInterface.java
@@ -1,0 +1,10 @@
+package com.alexandrialms.dao.interfaces;
+
+import com.alexandrialms.model.Loan;
+import java.sql.SQLException;
+import java.util.List;
+
+public interface LoanDAOInterface extends GenericDAO<Loan, Integer> {
+    List<Loan> findActiveLoans() throws SQLException;
+    List<Loan> findByUser(int userID) throws SQLException;
+}