+++ src/main/java/com/alexandrialms/model/CopyStatus.java
@@ -1,0 +1,9 @@
+package com.alexandrialms.model;
+
+public enum CopyStatus {
+    AVAILABLE,
+    BORROWED,
+    RESERVED,
+    DAMAGED,
+    LOST
+}