+++ src/main/java/com/alexandrialms/model/Password.java
@@ -1,7 +1,6 @@
 package com.alexandrialms.model;
 
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 
 public class Password {
     private int passwordID;