+++ src/main/java/com/alexandrialms/model/User.java
@@ -1,7 +1,6 @@
 package com.alexandrialms.model;
 
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 
 public class User {
     private int userID;