+++ src/main/java/com/alexandrialms/util/DBConnection.java
@@ -1,7 +1,6 @@
     package com.alexandrialms.util;
 
     import java.io.InputStream;
-    import java.io.InputStreamReader;
     import java.sql.Connection;
     import java.sql.DriverManager;
     import java.sql.SQLException;